// src/api/generated/models/transferStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */

export type TransferStatus = typeof TransferStatus[keyof typeof TransferStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransferStatus = {
  pending: 'pending',
  accepted: 'accepted',
  cancelled: 'cancelled',
} as const;